move docs to lib root
